Add context to share props between components
Add/create context in order to pass cartItemCount to ErrorPage.jsx. Due
to use of react router, ErrorPage.jsx and App.jsx are completely
separated components that do not share a common parent, so there was no
other way to do this. Previously had { cartItemCount, setItemCount } in
App.jsx but there was no way to pass those to ErrorPage.jsx. Now, the
cartItemCount can display properly in the nav of the error page.