Improve 'add to cart' popup display
Previously, if an item got added to the cart and the popup message
displayed, if another item got added to the cart before the first popup
completely faded out, the popup content for the new item was
just added inside the previous item's popup, so it would continue
fading out according to the first popup timer, or in some cases not
show at all.

Now, if a second item is added to the cart before a previously-added
item's popup completely fades out, the previous popup is replaced by a
new popup & timer for the second item.