Move cart & setCart to SharedProvider.jsx
Previously, if a user had items in their cart, and then clicks on a
broken link somewhere in the app that led to the error page rendered by
ErrorPage.jsx, the number of cart items shown in the nav would remain
accurate, but the actual cart content would be lost. Moving cart &
setCart to SharedProvider.jsx allows the cart content to persist
between ErrorPage.jsx and the other components.